+++ test-resources/eastwood/config/disable_unused_meta_on_macro.clj
@@ -1,0 +1,4 @@
+(disable-warning
+ {:linter :unused-meta-on-macro
+  :if-inside-macroexpansion-of '#{clojure.core/let}
+  :reason "Support a deftest"})