+++ changes.md
@@ -8,6 +8,8 @@
   * Related: the `disable-warnings` that Eastwood ships by default now prevent false positives against the [speced.def](https://github.com/nedap/speced.def) lib. 
 * Now the `:suspicious-test` linter can also be configured via the [`disable-warning`](https://github.com/jonase/eastwood#eastwood-config-files) mechanism.
   * Relatedly, a certain pattern of usage of the `clojure.test/are` macro now does not trigger a linter fault. 
+* Now the `:unused-meta-on-macro` linter can also be configured via the [`disable-warning`](https://github.com/jonase/eastwood#eastwood-config-files) mechanism.
+  * Closes https://github.com/jonase/eastwood/issues/303
 * If you invoke `eastwood.lint/eastwood` programatically, now a new key is offered: `:some-errors`, akin to `:some-warnings`.
   * This allows to distinguish why did Eastwood fail.